Move the path segments function to pydantic
And a bit of other stuff
